Remove PhantomData from Transit type
By associating the type T with the Transit struct, any lifetime
parameters associated with type T would become associated with the
Transit struct. This is bad because we would like to be able to
replicate the normal UDP socket behaviour. Removing this means that
Transit can send and receive different types all through the same
socket, and it will pick which type to serialize into based on type
inference.